Actualización de modulos de inventario, labor, configuraciónes, base y configuarciones del setting
