--- app_modul/templates/datos_finca.html
@@ -1,95 +1,0 @@
-{% extends 'base.html' %} 
-{% load static %} 
-{% block content %}
-
-<html>
-  <head>
-    <link rel="stylesheet" href="#" />
-    <title>Datos de la Finca</title>
-  </head>
-
-  <body>
-    <div class="Contenido">
-      <h3 class="tex">Información de la Finca</h3>
-
-      <div class="form">
-        <div class="dts">
-          <div class="mb-3">
-            <input
-              type="text"
-              class="form-control"
-              id="nit"
-              placeholder="Nit"
-            />
-          </div>
-
-          <div class="mb-3">
-            <input
-              type="text"
-              class="form-control"
-              id="nm-finca"
-              placeholder="Nombre de la Finca"
-            />
-          </div>
-
-          <div class="mb-3">
-            <input
-              type="text"
-              class="form-control"
-              id="nm-rpble"
-              placeholder="Nombre del Responsable"
-            />
-          </div>
-
-          <div class="mb-3">
-            <input
-              type="text"
-              class="form-control"
-              id="tel-rpble"
-              placeholder="Telefono del Responsable"
-            />
-          </div>
-
-          <div class="mb-3">
-            <input
-              type="text"
-              class="form-control"
-              id="direc"
-              placeholder="Dirección"
-            />
-          </div>
-        </div>
-      </div>
-
-      <div id="foto"></div>
-
-      <a
-        href="../Mas/index_mas.html"
-        class="btn btn-primary"
-        role="button"
-        id="btn_mas"
-        >Mas</a
-      >
-
-      <a
-        href="actualizar.html"
-        class="btn btn-primary"
-        role="button"
-        id="btn_actualizar"
-        >Actualizar</a
-      >
-
-      <a
-        href="../Ini_sesion/in_ses.html"
-        class="btn btn-primary"
-        role="button"
-        id="btn_cerrar"
-        >Cerrar Sesion</a
-      >
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
-  </body>
-</html>
-
-{% endblock %}
