+++ labor/templates/editarlabor.html
@@ -1,62 +1,73 @@
 {% extends 'base.html' %}
 {% load static %}
 {% block content %}
+<!DOCTYPE html>
 <html>
-<head>
-  
-</head>
-<table class="table">
-    <thead>
-      <tr>
-        <th>Descripcion</th>
-        <th>Lote</th>
-        <th>Fecha</th>
-        <th>Empleado</th>
-        <th>Estado</th>
-        <th>Opciones</th>
-      </tr>
-    </thead>
-    <tbody>
-      {% for list in listar_labor %}
-      <tr>
-        <td>{{ list.descripcion }}</td>
-        <td>{{ list.lote }}</td>
-        <td>{{ list.fecha_labor }}</td>
-        <td>{{ list.id_empleado.nombre }}</td>
-        <td>{{ list.id_estado.estado }}</td>
-        <td>
-          <a href="{% url 'editarlabor' list.id_labor %}"><button type="button" class="btn btn-info" data-bs-toggle="modal" data-bs-target="#edicionModal" id="btnEditar">Editar</button></a>
-        </td>
-      </tr>
-      {% endfor %}
-    </tbody>
-  </table>
+  <head>
+    <link rel="stylesheet" href="{% static 'css/tabla.css' %}">
+  </head>
+  <body>
+    <div class="gestion_tbla">
+      <div class="opciones">
+        <a href="{% url 'labor' %}"><button type="button" class="btn btn-secondary">Regresar</button></a>
+        <button type="button" class="btn btn-primary" data-bs-toggle="modal" 
+        data-bs-target="#edicionModal" data-id="1">Confirmar Edición</button>
+      </div>
 
-  <div class="modal fade" id="edicionModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-      <div class="modal-content">
-        <div class="modal-header">
-          <h5 class="modal-title" id="exampleModalLabel">New message</h5>
-        </div>
-        <div class="modal-body">
-          <form id="formulario" method="post">
-            {% csrf_token %}
-            {{data.form}}
-        
-
-           <div class="modal-footer">
-             <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-             <button type="submit" class="btn btn-primary">Guardar</button>
-          </form>
-        </div>
-        
+      <div class="table-container">
+        <div class="scrollable-table">
+          <table>
+            <thead>
+              <tr>
+                <th>Empleado</th>
+                <th>Labor</th>
+                <th>Lote</th>
+                <th>Fecha</th>
+                <th>Opciones</th>
+              </tr>
+            </thead>
+            <tbody>
+              {% for list in listar_labor %}
+              <tr class="lista">
+                <td>{{ list.id_empleado.nombre }}</td>
+                <td>{{ list.descripcion }}</td>
+                <td>{{ list.lote }}</td>
+                <td>{{ list.fecha_labor }}</td>
+                <td>
+                    <a href="{% url 'editarlabor' list.id_labor %}">
+                      <button type="button" class="btn btn-info">Editar</button>
+                    </a>
+                </td>
+              </tr>
+            {% endfor %}
+            </tbody>
+          </table>
         </div>
       </div>
     </div>
-  </div>
-  <h2>
-  <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#edicionModal" data-id="1">Confirmar Edición</button>
-  <a href="{% url 'labor' %}"><button type="button" class="btn btn-secondary">Volver</button></a>
-</body>
+
+    <div class="modal fade" id="edicionModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="exampleModalLabel">New message</h5>
+          </div>
+          <div class="modal-body">
+            <form id="formulario" method="post">
+              {% csrf_token %}
+              {{data.form}}
+          
+
+            <div class="modal-footer">
+              <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+              <button type="submit" class="btn btn-primary">Guardar</button>
+            </form>
+          </div>
+          
+          </div>
+        </div>
+      </div>
+    </div>
+  </body>
 </html>
 {% endblock %}
