+++ labor/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.http import JsonResponse
 from .forms import Laborform
-from . import models
+from gestion import models
 
 # Editar Labor
 def editarlabor(request, id_labor):
@@ -32,7 +32,7 @@ def obtenerempleado(request):
 
 # Listar Labores
 def labor(request):
-    listar_labor = models.Labor.objects.select_related('id_empleado', 'id_estado').all()
+    listar_labor = models.Labor.objects.select_related('id_empleado').all()
     return render(request, 'labor.html', {'listar_labor': listar_labor})
 
 # Agregar Registro de Labor
@@ -42,9 +42,6 @@ def agregaregistrolabor(request):
         lote = request.POST.get('lote', '')
         fecha_labor = request.POST.get('fecha_labor', '')
         id_empleado = request.POST.get('id_empleado', '')
-        id_estado = request.POST.get('id_estado', '')
-
-        estado = models.Estado.objects.get(id_estado=id_estado)
         empleado = models.Empleado.objects.get(id_empleado=id_empleado)
 
         form_data = {
@@ -52,7 +49,6 @@ def agregaregistrolabor(request):
             'lote': lote,
             'fecha_labor': fecha_labor,
             'id_empleado': empleado,
-            'id_estado': estado
         }
 
         form = Laborform(form_data)
